requirements.txt
bokeh
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
future==0.18.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
ignite==1.1.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
nbformat==4.4.0
numpy==1.16.0
opencv-python==4.2.0.34
pandas==0.23.4
Pillow==7.1.2
plotly==3.5.0
protobuf==3.12.2
pyparsing==2.3.1
python-dateutil==2.7.5
pytorch-ignite==0.3.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.10.1
six==1.12.0
tensorboard==1.12.2
tensorboardX==2.0
tensorflow==1.12.0
termcolor==1.1.0
torch==1.5.1
torchvision==0.6.1
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
wget==3.2
wrapt==1.12.1
